refactor: Avoid redefining version parsing twice in backend (bug 1879094)
