+++ api/src/shipit_api/admin/release.py
@@ -3,49 +3,16 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-import importlib
 import logging
-from functools import cache
 
 import requests
 from mozilla_version.gecko import FirefoxVersion
 
-from backend_common import get_products_config
 from shipit_api.common.config import SUPPORTED_FLAVORS
-from shipit_api.common.product import Product, get_key
+from shipit_api.common.product import Product
+from shipit_api.common.version import parse_version
 
 logger = logging.getLogger(__name__)
-
-
-@cache
-def _get_version_class_per_product_name():
-    version_class_per_product = {}
-    for product_name, product_config in get_products_config().items():
-        product_enum = Product[get_key(product_name)]
-        version_class_path = product_config["version-class"]
-        module_path = ".".join(version_class_path.split(".")[:-1])
-        class_name = version_class_path.split(".")[-1]
-        VersionClass = getattr(importlib.import_module(module_path), class_name, None)
-        version_class_per_product[product_enum] = VersionClass
-
-    return version_class_per_product
-
-
-def parse_version(product, version):
-    if isinstance(product, Product):
-        product_enum = product
-    else:
-        try:
-            product_enum = Product[get_key(product)]
-        except KeyError:
-            raise ValueError(f"Product {product} versions are not supported")
-
-    try:
-        VersionClass = _get_version_class_per_product_name()[product_enum]
-    except KeyError:
-        raise ValueError(f"Product {product} versions are not supported")
-
-    return VersionClass.parse(version)
 
 
 def is_rc(product, version, partial_updates):
